requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
beartype==0.20.2
billiard==4.2.1
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
crispy-tailwind==1.0.3
cryptography==43.0.0
cycler==0.12.1
daphne==4.1.2
Django==5.1
django-appconf==1.1.0
django-browser-reload==1.18.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-encrypted-model-fields==0.6.5
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-select2==8.4.0
django-simple-history==3.5.0
django-taggit==6.1.0
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
django-redis==5.2.0
dotenv==0.9.9
drf-yasg==1.21.8
et_xmlfile==2.0.0
Faker==37.1.0
fast_bitrix24==1.8.4
fonttools==4.56.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.26.0
hyperlink==21.0.0
icontract==2.7.1
idna==3.7
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.4
kiwisolver==1.4.8
kombu==5.4.2
MarkupSafe==2.1.5
matplotlib==3.10.1
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.4.3
narwhals==1.31.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pipfreeze==2.0.3
plotly==6.0.1
ply==3.11
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==1.7.1
pymemcache==4.0.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pysmi==1.4.4
pysnmp==6.2.5
pysnmpcrypto==0.0.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-memcached==1.62
python-telegram-bot==20.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
XlsxWriter==3.2.2
yarl==1.19.0
zope.interface==7.2
django-simpleui
qrcode[pil]
#django-qr-code
django-phonenumber-field